auth_service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6 